[Refactor] 변경된 UIFont 구조 적용
