+++ star/star/Source/Presentation/AppLaunch/AppLaunchViewController.swift
@@ -32,7 +32,7 @@ final class AppLaunchViewController: UIViewController {
     private lazy var titleLabel = UILabel().then {
         $0.text = "별을 이루는 당신의 시간, STAR와 함께 빛나세요"
         $0.textColor = .white
-        $0.font = Fonts.modalSectionOption
+        $0.font = UIFont.System.regular16
     }
     
     // MARK: - 생명주기 메서드
